fix: eliminate Edge Runtime warnings from instrumentation
Move Prisma-dependent onRequestError logic to a separate module
and dynamically import it behind a NEXT_RUNTIME check, preventing
the bundler from including Prisma in the Edge bundle.
